requirements.txt
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.1.0
fastjsonschema==2.16.2
Flask==2.2.2
flit_core==3.7.1
fonttools==4.37.4
fqdn==1.5.1
fsspec==2022.8.2
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
importlib==1.0.4
importlib-metadata==4.11.4
importlib-resources==5.9.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter-events==0.5.0
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.18.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
MouseInfo==0.1.3
nbclassic==0.4.8
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
notebook==6.4.12
notebook_shim==0.2.2
numpy==1.23.3
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==11.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyodbc==4.0.35
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytweening==1.0.4
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==303
pywinpty==2.0.10
pyxlsb==1.0.9
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
selenium==4.8.3
Send2Trash==1.8.0
setuptools==63.4.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
streamlit==1.21.0
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.3.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3==1.26.12
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.5
wikipedia==1.4.0
wincertstore==0.2
wsproto==1.2.0
XlsxWriter==3.0.6
zipp==3.8.1
